+++ integration_tests/nnbd_opted_in/test/opted_out_test.dart
@@ -1,0 +1,15 @@
+// Copyright (c) 2021, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// @dart=2.9
+
+import 'package:test/test.dart';
+
+import 'common/is_opted_out.dart';
+
+void main() {
+  test('unsound behavior', () async {
+    expect(isOptedOut, true);
+  });
+}